chore: creating image plugin grouping all image related functionality
